删掉重复文件
